add confirmation dialog for delete

+++ ufund-ui/angular/src/app/delete-need/delete-need.component.ts
@@ -1,6 +1,8 @@
 import { Component, Input } from '@angular/core';
 import { Need } from '../Need';
 import { CupboardService } from '../cupboard.service';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-delete-need',
@@ -11,22 +13,32 @@ export class DeleteNeedComponent {
   @Input() need?: Need;
   isAdmin: number = 1; // TODO: Implement once login is done
 
-  constructor(private cupboardService: CupboardService){}
+  constructor(private cupboardService: CupboardService, public dialog: MatDialog){}
   
   deleteNeed(): void {
-    if(!this.need) {
-      console.error('No need to delete');
-      return;
-    }
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      disableClose: false
+    });
+    dialogRef.componentInstance.confirmMessage = `Are you sure you want to delete ${this.need?.name}?`
 
-    const id = this.need.id;
-
-    this.cupboardService.deleteNeed(id).subscribe({
-      next: (response) => {
-        console.log('Sucessfully deleted need', response);
-      },
-      error: (error) => {
-        console.error('Error deleting need', error);
+    dialogRef.afterClosed().subscribe(result => {
+      console.log('The dialog was closed');
+      if (result) {
+        if(!this.need) {
+          console.error('No Need to delete');
+          return;
+        }
+    
+        const id = this.need.id;
+    
+        this.cupboardService.deleteNeed(id).subscribe({
+          next: (response) => {
+            console.log('Sucessfully deleted Need', response);
+          },
+          error: (error) => {
+            console.error('Error deleting Need', error);
+          }
+        });
       }
     });
   }
